+++ src/VideoEncoder/VideoEnc_FFV1.cpp
@@ -29,7 +29,7 @@ public:
 
 void ConfigFFV1::change_format(int sel)
 {
-	if (sel >= 0 && sel < std::size(codec->formats)) {
+	if (sel >= 0 && sel < (int)std::size(codec->formats)) {
 		codec->config->format = codec->formats[sel];
 		adjust_bits();
 		init_bits();
