+++ src/VideoEncoder/VideoEnc_ProRes.cpp
@@ -75,7 +75,7 @@ INT_PTR ConfigProres::DlgProc(UINT msg, WPARAM wParam, LPARAM lParam)
 
 void ConfigProres::change_format(int sel)
 {
-	if (sel >= 0 && sel < std::size(codec->formats)) {
+	if (sel >= 0 && sel < (int)std::size(codec->formats)) {
 		codec->config->format = codec->formats[sel];
 		init_profile();
 	}
